move dashboard files into subfolders.
