+++ src/components/dashboards/workshops/workshopsCarousel.module.css
